requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Babel
python-dotenv
pandas
joblib
Pillow
google-generativeai
scikit-learn
plotly
opencv-python
numpy
psutil
# For modern look, Bootstrap will be included via CDN in HTML